Updates
Al Salam Alaikom, 

I still need to read the remaining content of the file In Sha'a Allah. I stopped at line 93, and thus far I have the following questions. @naveed-ahmad could you please let me know

Are we using Bootstrap 4 or 3?
Which jQuery version, and which version for webpacker?
is rvm an option selected during the installation processes or is this a separate tool that needs to be set up individually?
Why does using rvm avoid potential problems? Could you please be more specific? if it is a known and complicated issue, then we can add a separate section to it? 

Thank you,