Update next page ref when jumping verses
Currently, when jumping to a verse in a surah, it won't update the ref to the next page. This means that if you jump from the first page to ayah 33 in a surah (next page would be 5 in this case), it won't update the next ref attribute to point to page 5. It will keep the page ref at 2. So, if you continue to scroll down and autoload the next page, it will load the ayahs from page 2.

To fix, whenever we use the jump to feature, we update the next page ref to point to next page according to the last verse retrieved.