refreshed readme.md
Step 8 at line 141, and step 10 at line 145 are duplicates. Step 9 might be out of place, and it it's missing the instruction.
I'm using the command line from a previous version of this readme file - commit id: 2540e1c

A previous version of this file had step 11 and step 12 in the reverse order. @naveed-ahmad , could you please confirm that step 11 and step 12 in this file are in the correct order? Thank you in advance.